requirements.txt
imagededup>=0.2.4
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
tqdm>=4.60.0
matplotlib>=3.3.0 